[tests] split bullet world tetsts into testCase and unittests
