+++ test/test_bullet_world_reasoning.py
@@ -1,12 +1,12 @@
 import time
 
-from test_bullet_world import BulletWorldTest
+from bullet_world_testcase import BulletWorldTestCase
 import pycram.bullet_world_reasoning as btr
 from pycram.pose import Pose
 from pycram.robot_descriptions import robot_description
 
 
-class TestBulletWorldReasoning(BulletWorldTest):
+class TestCaseBulletWorldReasoning(BulletWorldTestCase):
 
     def test_contact(self):
         self.milk.set_pose(Pose([1, 1, 1]))
